Cop to delete all block comments
Totally delete all text between =begin and =end tags, as well as
the tags themselves.

This cop was created while working on a legacy rails project which had
large chunks of code commented out using =begin and =end.

Do no confuse this cop with the Style/BlockComments cop which switches
=begin =end style comments to be # style comments. =begin comments
were popular before 2010, but have gone out of style because they
are more difficult for syntax highlighters to cope with in editors and
they are easy to miss when a programmer is simply scanning through
a file.